widen key constraint to 'any' for a standard map
